+++ src/soilmoisture.cpp
@@ -662,7 +662,7 @@ FrostComponent::updateLambdaRedux()
 
   for (size_t i_Layer = 0; i_Layer < vs_number_of_layers; i_Layer++) {
 
-    if (i_Layer < (int(floor((vm_FrostDepth / soilColumn[i_Layer].vs_LayerThickness) + 0.5)))) {
+    if (i_Layer < (size_t(floor((vm_FrostDepth / soilColumn[i_Layer].vs_LayerThickness) + 0.5)))) {
 
       // soil layer is frozen
       soilColumn[i_Layer].vs_SoilFrozen = true;
@@ -673,7 +673,7 @@ FrostComponent::updateLambdaRedux()
       }
     }
 
-    if (i_Layer < (int(floor((vm_ThawDepth / soilColumn[i_Layer].vs_LayerThickness) + 0.5)))) {
+    if (i_Layer < (size_t(floor((vm_ThawDepth / soilColumn[i_Layer].vs_LayerThickness) + 0.5)))) {
       // soil layer is thawing
 
       if (vm_ThawDepth < (double(i_Layer + 1) * soilColumn[i_Layer].vs_LayerThickness) && (vm_ThawDepth < vm_FrostDepth)) {
@@ -702,7 +702,7 @@ FrostComponent::updateLambdaRedux()
       vm_FrostDays = 0;
 
       vm_HydraulicConductivityRedux = centralParameterProvider.userSoilMoistureParameters.pm_HydraulicConductivityRedux;
-      for (int i_Layer = 0; i_Layer < vs_number_of_layers; i_Layer++) {
+      for (size_t i_Layer = 0; i_Layer < vs_number_of_layers; i_Layer++) {
         soilColumn[i_Layer].vs_SoilFrozen = false;
         vm_LambdaRedux[i_Layer] = 1.0;
       }
